Tip Calculator
Tip Calculator is used to calculate the Tip tht you wish to give to the person who served you well. Here this calculator need some parameter to say you how much you have to contribute or you have to pay for. Here are Some Required field that make it useful :

Total Bill: In this field the user need to add the Total amount which have to be pay as a bill.
Tip %: It include what percent of total bill you wish to pay as tip. The increment and decrement button help you to manage the value of Percent of Tip. If clicked for increment then it increment the value with 1 else if clicked for decrement the value of Tip% must decrease by 1 but not more than 100% and not less than zero.
Number of People: In This input field the user have to enter the number of people among whom the bill amount have to be divided. By this We will be able to find Tip given by per person and total amount have to be contributed per person.
It is interesting Project.